+++ vogs/src/compare_numbers.c
@@ -12,7 +12,7 @@
 
 #include "push_swap.h"
 
-int			compare(long num1, long num2)
+int			compare(int num1, int num2)
 {
 	if (num1 > num2)
 		return (1);
